@@ -1,19 +1,42 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net6.0</TargetFramework>
+		<Copyright>Copyright © 2019-2021 Mark Hazleton</Copyright>
+		<IsPackable>false</IsPackable>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+		<PackageReleaseNotes>Add Comments</PackageReleaseNotes>
+		<RepositoryType>GIT</RepositoryType>
+		<RepositoryUrl>https://github.com/markhazleton/SampleMvcCRUD</RepositoryUrl>
+		<PackageProjectUrl>https://github.com/markhazleton/SampleMvcCRUD</PackageProjectUrl>
+		<PackageIcon>ProjectMechanicsLogo.png</PackageIcon>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<Company>Mark Hazleton</Company>
+		<Authors>Mark Hazleton</Authors>
+		<Description>HttpClient Factory Mwh.Sample project</Description>
+		<ApplicationIcon>favicon.ico</ApplicationIcon>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Deterministic>false</Deterministic>
+		<SourceRevisionId>build$([System.DateTime]::UtcNow.ToString("yyyyMMddHHmmss"))</SourceRevisionId>
+		<VersionPrefix>5</VersionPrefix>
+		<VersionSuffix>$([System.DateTime]::UtcNow.ToString(yyMM)).$([System.DateTime]::UtcNow.ToString(ddHH)).$([System.DateTime]::UtcNow.ToString(mmss))</VersionSuffix>
+		<Version Condition=" '$(VersionSuffix)' == '' ">$(VersionPrefix).0.0.1</Version>
+		<Version Condition=" '$(VersionSuffix)' != '' ">$(VersionPrefix).$(VersionSuffix)</Version>
+		<GenerateDocumentationFile>True</GenerateDocumentationFile>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <None Remove="Clients\RestClientBase.cs~RF276f3a5b.TMP" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Remove="Clients\RestClientBase.cs~RF276f3a5b.TMP" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Mwh.Sample.Common\Mwh.Sample.Common.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\Mwh.Sample.Common\Mwh.Sample.Common.csproj" />
+	</ItemGroup>
 
 </Project>
