@@ -1,9 +1,7 @@
-﻿namespace Mwh.Sample.Common.Interfaces
+﻿namespace Mwh.Sample.Common.Interfaces;
+/// <summary>
+///
+/// </summary>
+public interface IEmployeeClient : IEmployeeService
 {
-    /// <summary>
-    ///
-    /// </summary>
-    public interface IEmployeeClient : IEmployeeService
-    {
-    }
 }
