@@ -1,12 +1,28 @@
-﻿namespace Mwh.Sample.Common.Clients
+﻿namespace Mwh.Sample.Common.Clients;
+/// <summary>
+/// REST Client Base Interface
+/// </summary>
+public interface IRestClientBase
 {
-    public interface IRestClientBase
-    {
-        void Dispose();
+    /// <summary>
+    /// Implement IDisposable
+    /// </summary>
+    void Dispose();
 
-        string AppName { get; set; }
-        string BaseAPIUrl { get; set; }
-        bool IsError { get; set; }
-        string Status { get; set; }
-    }
+    /// <summary>
+    /// Application Name added to header of all requests
+    /// </summary>
+    string AppName { get; set; }
+    /// <summary>
+    /// Base URL for all Requests
+    /// </summary>
+    string BaseAPIUrl { get; set; }
+    /// <summary>
+    /// Is Error
+    /// </summary>
+    bool IsError { get; set; }
+    /// <summary>
+    /// Status
+    /// </summary>
+    string Status { get; set; }
 }
