@@ -1,142 +1,133 @@
-﻿using Mwh.Sample.Common.Interfaces;
-using Mwh.Sample.Common.Models;
-using System;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Mwh.Sample.Common.Repositories
+﻿namespace Mwh.Sample.Common.Repositories;
+/// <summary>
+/// Class EmployeeService.
+/// Implements the <see cref="Mwh.Sample.Common.Repositories.IEmployeeService" />
+/// </summary>
+/// <seealso cref="Mwh.Sample.Common.Repositories.IEmployeeService" />
+public class EmployeeService : IEmployeeService
 {
     /// <summary>
-    /// Class EmployeeService.
-    /// Implements the <see cref="Mwh.Sample.Common.Repositories.IEmployeeService" />
+    /// The employee repository
     /// </summary>
-    /// <seealso cref="Mwh.Sample.Common.Repositories.IEmployeeService" />
-    public class EmployeeService : IEmployeeService
+    private readonly IEmployeeRepository _employeeRepository;
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="EmployeeService"/> class.
+    /// </summary>
+    /// <param name="employeeRepository">The employee repository.</param>
+    public EmployeeService(IEmployeeRepository employeeRepository)
     {
-        /// <summary>
-        /// The employee repository
-        /// </summary>
-        private readonly IEmployeeRepository _employeeRepository;
+        _employeeRepository = employeeRepository;
+    }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="EmployeeService"/> class.
-        /// </summary>
-        /// <param name="employeeRepository">The employee repository.</param>
-        public EmployeeService(IEmployeeRepository employeeRepository)
+    /// <summary>
+    /// Lists the asynchronous.
+    /// </summary>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>Task&lt;IEnumerable&lt;EmployeeModel&gt;&gt;.</returns>
+    public Task<IEnumerable<EmployeeModel>> GetAsync(CancellationToken token)
+    {
+        return _employeeRepository.ListAsync(token);
+    }
+
+    /// <summary>
+    /// Finds the by identifier asynchronous.
+    /// </summary>
+    /// <param name="id">The identifier.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>Task&lt;EmployeeModel&gt;.</returns>
+    public Task<EmployeeModel> FindByIdAsync(int id, CancellationToken token)
+    {
+        return _employeeRepository.FindByIdAsync(id, token);
+    }
+
+    /// <summary>
+    /// save as an asynchronous operation.
+    /// </summary>
+    /// <param name="employee">The employee.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>EmployeeResponse.</returns>
+    public async Task<EmployeeResponse> SaveAsync(EmployeeModel employee, CancellationToken token)
+    {
+        if (employee == null) return new EmployeeResponse("Employee is null");
+        try
         {
-            _employeeRepository = employeeRepository;
+            var response = await _employeeRepository.AddAsync(employee, token).ConfigureAwait(true);
+
+            if (response == null) return new EmployeeResponse("Repository Response was null");
+
+            return new EmployeeResponse(response);
         }
-
-        /// <summary>
-        /// Lists the asynchronous.
-        /// </summary>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task&lt;IEnumerable&lt;EmployeeModel&gt;&gt;.</returns>
-        public Task<IEnumerable<EmployeeModel>> GetAsync(CancellationToken token)
+        catch (Exception ex)
         {
-            return _employeeRepository.ListAsync(token);
+            // Do some logging stuff
+            return new EmployeeResponse($"An error occurred when saving the employee: {ex.Message}");
         }
+    }
 
-        /// <summary>
-        /// Finds the by identifier asynchronous.
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task&lt;EmployeeModel&gt;.</returns>
-        public Task<EmployeeModel> FindByIdAsync(int id, CancellationToken token)
+    /// <summary>
+    /// update as an asynchronous operation.
+    /// </summary>
+    /// <param name="id">The identifier.</param>
+    /// <param name="employee">The employee.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>EmployeeResponse.</returns>
+    public async Task<EmployeeResponse> UpdateAsync(int id, EmployeeModel employee, CancellationToken token)
+    {
+        if (employee == null)
+            return new EmployeeResponse("Employee is null.");
+
+        if (employee.id != id)
+            return new EmployeeResponse($"Mismatch in id({id}) && id({employee.id}).");
+
+        var existingEmployee = await _employeeRepository.FindByIdAsync(id, token).ConfigureAwait(true);
+        if (existingEmployee == null)
+            return new EmployeeResponse("Employee not found.");
+
+        try
         {
-            return _employeeRepository.FindByIdAsync(id, token);
+            var response = await _employeeRepository.UpdateAsync(employee, token).ConfigureAwait(true);
+
+            if (response == null) return new EmployeeResponse("Repository Response was null");
+
+            return new EmployeeResponse(response);
         }
-
-        /// <summary>
-        /// save as an asynchronous operation.
-        /// </summary>
-        /// <param name="employee">The employee.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>EmployeeResponse.</returns>
-        public async Task<EmployeeResponse> SaveAsync(EmployeeModel employee, CancellationToken token)
+        catch (Exception ex)
         {
-            if (employee == null) return new EmployeeResponse("Employee is null");
-            try
-            {
-                var response = await _employeeRepository.AddAsync(employee, token).ConfigureAwait(true);
-
-                if (response == null) return new EmployeeResponse("Repository Response was null");
-
-                return new EmployeeResponse(response);
-            }
-            catch (Exception ex)
-            {
-                // Do some logging stuff
-                return new EmployeeResponse($"An error occurred when saving the employee: {ex.Message}");
-            }
+            // Do some logging stuff
+            return new EmployeeResponse($"An error occurred when updating the employee: {ex.Message}");
         }
+    }
 
-        /// <summary>
-        /// update as an asynchronous operation.
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <param name="employee">The employee.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>EmployeeResponse.</returns>
-        public async Task<EmployeeResponse> UpdateAsync(int id, EmployeeModel employee, CancellationToken token)
+    /// <summary>
+    /// delete as an asynchronous operation.
+    /// </summary>
+    /// <param name="id">The identifier.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>EmployeeResponse.</returns>
+    public async Task<EmployeeResponse> DeleteAsync(int id, CancellationToken token)
+    {
+        var existingEmployee = await _employeeRepository.FindByIdAsync(id, token).ConfigureAwait(true);
+
+        if (existingEmployee == null)
+            return new EmployeeResponse("Employee not found.");
+
+        if (!existingEmployee.IsValid)
+            return new EmployeeResponse("Employee not found.");
+
+        try
         {
-            if (employee == null)
-                return new EmployeeResponse("Employee is null.");
+            var response = await _employeeRepository.RemoveAsync(existingEmployee, token).ConfigureAwait(true);
 
-            if (employee.id != id)
-                return new EmployeeResponse($"Mismatch in id({id}) && id({employee.id}).");
+            if (response)
+                existingEmployee.id = 0;
 
-            var existingEmployee = await _employeeRepository.FindByIdAsync(id, token).ConfigureAwait(true);
-            if (existingEmployee == null)
-                return new EmployeeResponse("Employee not found.");
-
-            try
-            {
-                var response = await _employeeRepository.UpdateAsync(employee, token).ConfigureAwait(true);
-
-                if (response == null) return new EmployeeResponse("Repository Response was null");
-
-                return new EmployeeResponse(response);
-            }
-            catch (Exception ex)
-            {
-                // Do some logging stuff
-                return new EmployeeResponse($"An error occurred when updating the employee: {ex.Message}");
-            }
+            return new EmployeeResponse(existingEmployee);
         }
-
-        /// <summary>
-        /// delete as an asynchronous operation.
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>EmployeeResponse.</returns>
-        public async Task<EmployeeResponse> DeleteAsync(int id, CancellationToken token)
+        catch (Exception ex)
         {
-            var existingEmployee = await _employeeRepository.FindByIdAsync(id, token).ConfigureAwait(true);
-
-            if (existingEmployee == null)
-                return new EmployeeResponse("Employee not found.");
-
-            if (!existingEmployee.IsValid)
-                return new EmployeeResponse("Employee not found.");
-
-            try
-            {
-                var response = await _employeeRepository.RemoveAsync(existingEmployee, token).ConfigureAwait(true);
-
-                if (response)
-                    existingEmployee.id = 0;
-
-                return new EmployeeResponse(existingEmployee);
-            }
-            catch (Exception ex)
-            {
-                // Do some logging stuff
-                return new EmployeeResponse($"An error occurred when deleting the employee: {ex.Message}");
-            }
+            // Do some logging stuff
+            return new EmployeeResponse($"An error occurred when deleting the employee: {ex.Message}");
         }
     }
 }
