@@ -1,85 +1,78 @@
-﻿using Mwh.Sample.Common.Interfaces;
-using Mwh.Sample.Common.Models;
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
+﻿namespace Mwh.Sample.Common.Repositories;
 
-namespace Mwh.Sample.Common.Repositories
+/// <summary>
+/// Class EmployeeRepository.
+/// Implements the <see cref="Repositories.IEmployeeRepository" />
+/// </summary>
+/// <seealso cref="Mwh.Sample.Common.Repositories.IEmployeeRepository" />
+public class EmployeeRepository : IEmployeeRepository
 {
     /// <summary>
-    /// Class EmployeeRepository.
-    /// Implements the <see cref="Mwh.Sample.Common.Repositories.IEmployeeRepository" />
+    /// The emp
     /// </summary>
-    /// <seealso cref="Mwh.Sample.Common.Repositories.IEmployeeRepository" />
-    public class EmployeeRepository : IEmployeeRepository
+    private readonly IEmployeeDB _emp;
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="EmployeeRepository"/> class.
+    /// </summary>
+    /// <param name="employeeDB">The employee database.</param>
+    public EmployeeRepository(IEmployeeDB employeeDB) { _emp = employeeDB; }
+
+    /// <summary>
+    /// add as an asynchronous operation.
+    /// </summary>
+    /// <param name="employee">The employee.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    public Task<EmployeeModel> AddAsync(EmployeeModel employee, CancellationToken token)
     {
-        /// <summary>
-        /// The emp
-        /// </summary>
-        private readonly IEmployeeDB _emp;
+        token.ThrowIfCancellationRequested();
+        return Task.Run(() => _emp.Update(employee));
+    }
 
-        /// <summary>
-        /// Initializes a new instance of the <see cref="EmployeeRepository"/> class.
-        /// </summary>
-        /// <param name="employeeDB">The employee database.</param>
-        public EmployeeRepository(IEmployeeDB employeeDB) { _emp = employeeDB; }
+    /// <summary>
+    /// Finds the by identifier asynchronous.
+    /// </summary>
+    /// <param name="id">The identifier.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>Task&lt;EmployeeModel&gt;.</returns>
+    public Task<EmployeeModel> FindByIdAsync(int id, CancellationToken token)
+    {
+        token.ThrowIfCancellationRequested();
+        return Task.Run(() => _emp.Employee(id));
+    }
 
-        /// <summary>
-        /// add as an asynchronous operation.
-        /// </summary>
-        /// <param name="employee">The employee.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        public Task<EmployeeModel> AddAsync(EmployeeModel employee, CancellationToken token)
-        {
-            token.ThrowIfCancellationRequested();
-            return Task.Run(() => _emp.Update(employee));
-        }
+    /// <summary>
+    /// Lists the asynchronous.
+    /// </summary>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>Task&lt;IEnumerable&lt;EmployeeModel&gt;&gt;.</returns>
+    public Task<IEnumerable<EmployeeModel>> ListAsync(CancellationToken token)
+    {
+        token.ThrowIfCancellationRequested();
+        return Task.Run(() => (IEnumerable<EmployeeModel>)_emp.EmployeeCollection());
+    }
 
-        /// <summary>
-        /// Finds the by identifier asynchronous.
-        /// </summary>
-        /// <param name="id">The identifier.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task&lt;EmployeeModel&gt;.</returns>
-        public Task<EmployeeModel> FindByIdAsync(int id, CancellationToken token)
-        {
-            token.ThrowIfCancellationRequested();
-            return Task.Run(() => _emp.Employee(id));
-        }
+    /// <summary>
+    /// Removes the specified employee.
+    /// </summary>
+    /// <param name="employee">The employee.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>Task&lt;System.Boolean&gt;.</returns>
+    public Task<bool> RemoveAsync(EmployeeModel employee, CancellationToken token)
+    {
+        token.ThrowIfCancellationRequested();
+        return Task.Run(() => _emp.Delete(employee.id));
+    }
 
-        /// <summary>
-        /// Lists the asynchronous.
-        /// </summary>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task&lt;IEnumerable&lt;EmployeeModel&gt;&gt;.</returns>
-        public Task<IEnumerable<EmployeeModel>> ListAsync(CancellationToken token)
-        {
-            token.ThrowIfCancellationRequested();
-            return Task.Run(() => (IEnumerable<EmployeeModel>)_emp.EmployeeCollection());
-        }
-
-        /// <summary>
-        /// Removes the specified employee.
-        /// </summary>
-        /// <param name="employee">The employee.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task&lt;System.Boolean&gt;.</returns>
-        public Task<bool> RemoveAsync(EmployeeModel employee, CancellationToken token)
-        {
-            token.ThrowIfCancellationRequested();
-            return Task.Run(() => _emp.Delete(employee.id));
-        }
-
-        /// <summary>
-        /// Updates the specified employee.
-        /// </summary>
-        /// <param name="employee">The employee.</param>
-        /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-        /// <returns>Task&lt;EmployeeModel&gt;.</returns>
-        public Task<EmployeeModel> UpdateAsync(EmployeeModel employee, CancellationToken token)
-        {
-            token.ThrowIfCancellationRequested();
-            return Task.Run(() => _emp.Update(employee));
-        }
+    /// <summary>
+    /// Updates the specified employee.
+    /// </summary>
+    /// <param name="employee">The employee.</param>
+    /// <param name="token">The cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>Task&lt;EmployeeModel&gt;.</returns>
+    public Task<EmployeeModel> UpdateAsync(EmployeeModel employee, CancellationToken token)
+    {
+        token.ThrowIfCancellationRequested();
+        return Task.Run(() => _emp.Update(employee));
     }
 }
