@@ -1,51 +1,48 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿namespace Mwh.Sample.Core.Data.Models;
 
-namespace Mwh.Sample.Core.Data.Models
+public class Employee
 {
-    public class Employee
+    /// <summary>
+    /// Constructor
+    /// </summary>
+    public Employee()
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        public Employee()
-        {
-        }
-
-        /// <summary>
-        /// Gets or sets the age.
-        /// </summary>
-        /// <value>The age.</value>
-        public int Age { get; set; }
-
-        /// <summary>
-        /// Gets or sets the country.
-        /// </summary>
-        /// <value>The country.</value>
-        public string Country { get; set; }
-
-        /// <summary>
-        /// Gets or sets the department.
-        /// </summary>
-        /// <value>The department.</value>
-        public int DepartmentId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the employee identifier.
-        /// </summary>
-        /// <value>The employee identifier.</value>
-        [Key]
-        public int Id { get; set; }
-
-        /// <summary>
-        /// Gets or sets the name.
-        /// </summary>
-        /// <value>The name.</value>
-        public string Name { get; set; }
-
-        /// <summary>
-        /// Gets or sets the state.
-        /// </summary>
-        /// <value>The state.</value>
-        public string State { get; set; }
     }
+
+    /// <summary>
+    /// Gets or sets the age.
+    /// </summary>
+    /// <value>The age.</value>
+    public int Age { get; set; }
+
+    /// <summary>
+    /// Gets or sets the country.
+    /// </summary>
+    /// <value>The country.</value>
+    public string Country { get; set; }
+
+    /// <summary>
+    /// Gets or sets the department.
+    /// </summary>
+    /// <value>The department.</value>
+    public int DepartmentId { get; set; }
+
+    /// <summary>
+    /// Gets or sets the employee identifier.
+    /// </summary>
+    /// <value>The employee identifier.</value>
+    [Key]
+    public int Id { get; set; }
+
+    /// <summary>
+    /// Gets or sets the name.
+    /// </summary>
+    /// <value>The name.</value>
+    public string Name { get; set; }
+
+    /// <summary>
+    /// Gets or sets the state.
+    /// </summary>
+    /// <value>The state.</value>
+    public string State { get; set; }
 }
