@@ -1,31 +1,28 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿namespace Mwh.Sample.Core.Data.Models;
 
-namespace Mwh.Sample.Core.Data.Models
+public class EmployeeContext : DbContext
 {
-    public class EmployeeContext : DbContext
+    public EmployeeContext()
     {
-        public EmployeeContext()
-        {
-        }
-
-        public EmployeeContext(DbContextOptions options) : base(options)
-        {
-            ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-                optionsBuilder.UseInMemoryDatabase("DefaultDatabase");
-            }
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            // modelBuilder.Entity<Employee>().ToTable("Employees");
-        }
-
-        public DbSet<Employee> Employees { get; set; }
     }
+
+    public EmployeeContext(DbContextOptions options) : base(options)
+    {
+        ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
+    }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        if (!optionsBuilder.IsConfigured)
+        {
+            optionsBuilder.UseInMemoryDatabase("DefaultDatabase");
+        }
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        // modelBuilder.Entity<Employee>().ToTable("Employees");
+    }
+
+    public DbSet<Employee> Employees { get; set; }
 }
